tools/PocketTrusteeCLI/requirement.txt
rich
cmd2
pyserial
pillow
cbor2
pycryptodome
qrcode[pil]